+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - [#3511](https://github.com/influxdb/influxdb/issues/3511): Sending a large number of tag causes panic
 - [#3288](https://github.com/influxdb/influxdb/issues/3288): Run go fuzz on the line-protocol input
 - [#3545](https://github.com/influxdb/influxdb/issues/3545): Fix parsing string fields with newlines
+- [#3579](https://github.com/influxdb/influxdb/issues/3579): Revert breaking change to `client.NewClient` function
 
 ## v0.9.2 [2015-07-24]